remove planner.py
