+++ source/jormungandr/jormungandr/pt_planners/kraken.py
@@ -41,9 +41,6 @@ class Kraken(ZmqSocket, AbstractPtPlanner):
     def name(self):
         return self.instance_name
 
-    def stop_points_nearby(self, place, distance, access_points=False):
-        raise NotImplementedError("Too bad, you cannot ask kraken for places nearby :)")
-
     def journeys(self, origins, destinations, datetime, clockwise, journey_parameters, bike_in_pt, request_id):
         req = utils.create_journeys_request(
             origins, destinations, datetime, clockwise, journey_parameters, bike_in_pt
